autho/dev_requirements.txt
-r requirements.txt
django-extensions==1.6.7
py==1.4.31
pytest==2.9.2
pytest-django==2.9.1
six==1.10.0
